Several renamings; 'couple' -> 'resource';
 * The monad transformer formerly known as 'CoupleT' is now known as 'Resource'. This is inspired by Gabriel Gonzales's post [The Resource Monad](http://www.haskellforall.com/2013/06/the-resource-applicative.html), which describes a similar constrcut.
 * Accordingly, 'couple' has been renamed to 'resource', and 'decouple' has been renamed to 'acquire'.
 * The 'Resource' monad now lives in the module 'Data.Resource', as opposed to the longer 'Control.Monad.Trans.Couple'.
 * The 'DecoupleT' monad transformer has also been renamed, again inspired by Gabriel Gonzales. The [pipes-safe](http://hackage.haskell.org/package/pipes-safe) package has a very similar monad, which is also called 'SafeT'.

After renaming everything in this package several times, I think these names will be final. 'Resource' was always the obvious name for that monad, but I hesitated to call it Resource(T) because "ResourceT" is already used by Michael Snoyman for a monad similar to "SafeT", for which I also struggled to find a good name.