Renamed "Control.Monad.Interface.Safe.ReleaseKey" to "Control.Monad.Interface.Safe.Internal", and moved MonadSafe and associated operations there.
The nice thing is about this is that the SafeT part of the package and the Resource part of the package are almost completely separate from each other. All the glue between them is contained in the "Control.Monad.Interface.Safe" module (which also re-exports some of "Control.Monad.Interface.Safe.Internal"). In fact that is the only thing that module contains.