Added 'register'' and 'release'' operations to the 'MonadSafe' interface, as well as the 'resource'' constructor for 'Resource'.
It is now possible to specify that a different finalizer be run if the main code threw an exception as opposed to finishing successfully. This is useful in some instances, and makes it possible to model a 'ForeignPtr' as a 'Resource', for example. The behaviour of the non-primed versions of the above operations remains unchanged.