Added 'Finalizers' datatype to simplify the implementation of 'Resource'.
This in some way undoes some of the uncoupling achieved in the previous commit, but in other ways furthers it. It would still be possible to split the 'Resource' and 'SafeT'/'MonadSafe' parts, but now the latter would be firmly dependent on the former. But that's okay, it actually makes things simpler, and it means there is an obvious place to put the "glue": in the 'SafeT'/'MonadSafe' part.

The 'Finalizers' datatype makes several things explicit that were probably not clear before (the code is still mostly undocumentated). I think it also satisfactorily solves the 'unsafeAcquire' problem, such that the 'Resource' part of the library would actually be a viable library in its own right (even without the 'acquire' operation from the 'MonadSafe' interface). However I'm still going to keep them together.